kernel: 5/25-26 batch — sockets-as-fds + OP_POLL + style + atomics
Six logical workstreams accumulated since f813bec, squashed into one
commit because file-level changes overlap. Each has a memory entry
with the full context. Build 0x6A1577FB clean ReleaseSafe.

debug/kstack_protect: ripped. The per-page-protect-on-free tripwire
created recycled-slot R/O crashes; idea preserved in memory, file +
call sites (idt/lifecycle/sched/smp/desktop/kasan_inject) deleted.

proc/sched: silent-corruption hunt resolved. The mirror-flip detector
was false-firing during blockOn-yield's unbracketed window between two
consecutive call-pushes (sched.zig:1383 → 1578). scheduling_out_pid
bracket now set for ALL yields. 8-day hunt, 32 destroys post-fix with
zero mirror-flips.

proc/signals + proc/process: POSIX exit_group / killThreadGroup.
Fatal default-action signals (SIGINT, etc.) kill ALL peers in the
same tgid via new process.killThreadGroup, called from
signals.applyDefault. Fixes mtswap-style multi-thread leak where ^C
only reaped the leader and workers kept holding all memory.

cpu/iouring + cpu/fdpoll: OP_POLL landed. New src/cpu/fdpoll.zig
(~200 LOC: predicate + waiter registry) + iouring OP_POLL opcode.
Covers pipes + console fds + TCP fds (after sockets-as-fds below).

proc + fs + cpu/syscall + lib + app: sockets-as-fds. TCP conns +
listeners migrated into the per-process fd_table. process.FsType
gained .tcp_sock=7 + .tcp_listener=8; vfs.close/read/write got new
arms; cpu/syscall/common new allocSocketFd/resolveSocketFd helpers;
cpu/syscall/net handlers refactored to take/return fds; sys#73
(tcp_close) + sys#76 (tcp_unlisten) removed in favor of close(fd).
OP_POLL now covers TCP uniformly via fdpoll's tcp_sock arm.
Validated end-to-end via httpd (117 close() calls, 0 E_BADF).

docs/STYLE.md + canonical structs + wire-format extern structs:
house-style guide adopted. Field-access tags (p:lock)/(a)/(c) on
long-lived shared structs (PCB, pmm.Region, NVMe Controller +
NvmeWaiter, TcpConn/TcpListener, FileDesc, LazyRegion). Untagged
fields are owner-pid / owner-CPU implicit default. Plus comptime
@sizeOf + @offsetOf asserts directly under every wire-format extern
struct: NVMe SqEntry/CqEntry/DsmRange, virtio VirtqDesc, ext2
Superblock/BlockGroupDescriptor/Inode/DirEntry, fat32 BPB/DirEntry,
e1000 + i225 RxDesc/TxDesc.

Atomic-access audit on top of the new tags: zig-osdev-reviewer used
the (a) annotations as ground truth and surfaced 7 inconsistencies
between the declared discipline and actual access — the methodology
test paid off immediately. Fixed:
  * cpu/syscall + cpu/idt + cpu/syscall/proc: 3 plain reads of
    pending_signals → @atomicLoad. Sender does atomic OR; reader's
    plain load could miss the bit (delayed signal delivery class).
  * driver/nvme: 7 plain accesses of async_mode (incl. IRQ handler
    at line 381) → @atomicLoad/@atomicStore. The boot-time false→
    true flip could leave AP stuck on sync path silently.
  * proc/sched: 36 plain stores of wait_kind / wait_target
    converted via 3 new inline helpers (setWait, clearWait,
    waitsOn). This is the lost-wake class — exact family that
    drove the 8-day silent-corruption hunt and the futex lost-wake
    fix. wake-helper fan-out loops (wakeMutexWaiters,
    wakeSwapEvictWaiters, wakeIoUring{Worker,Cq}) and
    virtio_gpu's IRQ-context wait_kind scan also converted.
  * cpu/syscall/proc + proc/sched: 6 wake_tick plain reads/writes
    → @atomic*. Cross-CPU stale-read could delay sleep / gpu_io
    wake.

Tag corrections: wake_tick (p:rq.lock) → (a) (sysSleep + sched +
virtio_gpu IRQ all touch it); NvmeWaiter.active/gen (p:io_lock) →
(a) (allocCid uses CAS, not io_lock); NvmeWaiter.success /
status_code / wake / sq_slot annotated "doorbell-fenced" since
submitter writes before the SQ-doorbell MMIO write and reapCq
reads under cq_lock — the doorbell acts as the release fence.

+++ app/httpd.zig
@@ -154,7 +154,7 @@ fn jsonResult() []const u8 {
 /// first line (`GET /path HTTP/1.x`), so once we see `\r\n\r\n` we stop.
 /// Also bails immediately when `should_quit` flips — so a client that
 /// connected then went silent doesn't extend Ctrl+C latency by 3 s.
-fn readRequest(conn: u8, buf: []u8) ?usize {
+fn readRequest(conn: u32, buf: []u8) ?usize {
     var total: usize = 0;
     var iters: u32 = 0;
     while (iters < READ_TIMEOUT_ITERS) : (iters += 1) {
@@ -215,22 +215,22 @@ fn isSafePath(p: []const u8) bool {
     return true;
 }
 
-fn sendStatus(conn: u8, line: []const u8) void {
+fn sendStatus(conn: u32, line: []const u8) void {
     _ = libc.tcpSend(conn, "HTTP/1.0 ");
     _ = libc.tcpSend(conn, line);
     _ = libc.tcpSend(conn, "\r\nConnection: close\r\nContent-Length: 0\r\n\r\n");
 }
 
-fn sendNotFound(conn: u8) void {
+fn sendNotFound(conn: u32) void {
     sendStatus(conn, "404 Not Found");
 }
 
-fn sendBadRequest(conn: u8) void {
+fn sendBadRequest(conn: u32) void {
     sendStatus(conn, "400 Bad Request");
 }
 
 /// Send a JSON body with HTTP/1.0 200 + Content-Length + close.
-fn sendJsonResponse(conn: u8, body: []const u8) void {
+fn sendJsonResponse(conn: u32, body: []const u8) void {
     var len_buf: [16]u8 = undefined;
     const len_str = formatU32(&len_buf, @intCast(body.len));
     _ = libc.tcpSend(conn, "HTTP/1.0 200 OK\r\nContent-Type: application/json\r\nCache-Control: no-store\r\nAccess-Control-Allow-Origin: *\r\nContent-Length: ");
@@ -242,7 +242,7 @@ fn sendJsonResponse(conn: u8, body: []const u8) void {
 /// Plain-text variant of sendJsonResponse for endpoints that don't benefit
 /// from JSON wrapping (e.g. raw log dumps that the client renders into a
 /// <pre>). Same Content-Length / no-store / CORS headers.
-fn sendTextResponse(conn: u8, body: []const u8) void {
+fn sendTextResponse(conn: u32, body: []const u8) void {
     var len_buf: [16]u8 = undefined;
     const len_str = formatU32(&len_buf, @intCast(body.len));
     _ = libc.tcpSend(conn, "HTTP/1.0 200 OK\r\nContent-Type: text/plain; charset=utf-8\r\nCache-Control: no-store\r\nAccess-Control-Allow-Origin: *\r\nContent-Length: ");
@@ -399,7 +399,7 @@ fn apiAll() []const u8 {
 /// Dispatch /api/* routes. `route` is the path segment after /api/.
 /// Returns true if a route was matched (and a response sent), false to fall
 /// through to file serving.
-fn handleApi(conn: u8, route: []const u8) bool {
+fn handleApi(conn: u32, route: []const u8) bool {
     if (equals(route, "uptime")) {
         sendJsonResponse(conn, apiUptime());
         return true;
@@ -454,7 +454,7 @@ fn apiLog() []const u8 {
     return json_buf[0..json_len];
 }
 
-fn serveFile(conn: u8, doc_path: []const u8, ctype: []const u8) void {
+fn serveFile(conn: u32, doc_path: []const u8, ctype: []const u8) void {
     const size = libc.fsize(doc_path) orelse {
         sendNotFound(conn);
         return;
@@ -490,7 +490,7 @@ fn serveFile(conn: u8, doc_path: []const u8, ctype: []const u8) void {
     _ = libc.tcpSend(conn, buf[0..total]);
 }
 
-fn handleConn(conn: u8) void {
+fn handleConn(conn: u32) void {
     var req_buf: [REQUEST_BUF]u8 = undefined;
     const req_len = readRequest(conn, &req_buf) orelse {
         sendBadRequest(conn);
